Support impulsive attitude maneuvers
Repeated epochs now represent a discontinuity rather than being
rejected, matching the orbit and generic trajectories. An impulsive
attitude maneuver holds its pre- and post-maneuver state at the same
instant, and a query there returns the most recently added state, so
interpolation is right-continuous instead of dividing by a
zero-length interval and returning NaN.

Lagrange windowing moves to the shared compute_lagrange_window,
which confines the window to the run of strictly increasing epochs
containing the bracket. No polynomial can represent a jump, so the
window shrinks rather than fitting across one.

Split state_providers into a module directory, one file per provider
family, with every trait re-exported so existing paths resolve
unchanged. Document why slerp is the default interpolation method.
